Remove repeating sections from LMC and Adversary weapons
This will cause all lore master characters and adversaries to lose
their weapons, but unfortunately there is no way around that.

This change is necessary, because there is otherwise no way to make the
weapon rolls work, and I’ve now been contacted a bunch of times about
this.

So, we move forward begrudgingly.